+++ packages/figma/src/codegen/targets/react/component/handlers/divider.ts
@@ -4,16 +4,22 @@ import * as metadata from "@/entities/data/__generated__/component-sets";
 import { match } from "ts-pattern";
 import { createSeedReactElement } from "../../element-factories";
 import type { ComponentHandlerDeps } from "../deps.interface";
+import { camelCase } from "change-case";
 
 export const createDividerHandler = (_ctx: ComponentHandlerDeps) =>
   defineComponentHandler<DividerProperties>(
     metadata.divider.key,
     ({ componentProperties: props }) => {
       const { color } = match(props.Tone.value)
-        .with("Neutral", () => ({ color: "stroke.neutral" }))
         .with("Neutral Muted", () => ({ color: "stroke.neutralMuted" }))
+        .with("Neutral Subtle", () => ({ color: "stroke.neutralSubtle" }))
         .exhaustive();
 
-      return createSeedReactElement("Divider", { color });
+      const commonProps = {
+        color,
+        orientation: camelCase(props.Orientation.value),
+      };
+
+      return createSeedReactElement("Divider", commonProps, undefined);
     },
   );