+++ packages/figma/src/codegen/targets/react/component/index.ts
@@ -14,6 +14,7 @@ import { createBadgeHandler } from "./handlers/badge";
 import { createBottomSheetHandler } from "./handlers/bottom-sheet";
 import { createCalloutHandler } from "./handlers/callout";
 import { createCheckboxHandler } from "./handlers/checkbox";
+import { createCheckmarkHandler } from "@/codegen/targets/react/component/handlers/checkmark";
 import { createChipHandler } from "./handlers/chip";
 import { createContextualFloatingButtonHandler } from "./handlers/contextual-floating-button";
 import { createDividerHandler } from "./handlers/divider";
@@ -21,20 +22,24 @@ import { createErrorStateHandler } from "./handlers/error-state";
 import { createFloatingActionButtonHandler } from "./handlers/floating-action-button";
 import { createHelpBubbleHandler } from "./handlers/help-bubble";
 import { createIdentityPlaceholderHandler } from "./handlers/identity-placeholder";
-import { createPageBannerHandler } from "./handlers/page-banner";
 import { createMannerTempBadgeHandler } from "./handlers/manner-temp-badge";
 import { createMannerTempHandler } from "./handlers/manner-temp";
 import { createMenuSheetHandler } from "./handlers/menu-sheet";
 import { createMultilineTextFieldHandler } from "./handlers/multiline-text-field";
+import { createPageBannerHandler } from "./handlers/page-banner";
 import { createProgressCircleHandler } from "./handlers/progress-circle";
+import { createRadioMarkHandler } from "@/codegen/targets/react/component/handlers/radio-mark";
+import { createRadioGroupItemHandler } from "@/codegen/targets/react/component/handlers/radio-group";
 import { createReactionButtonHandler } from "./handlers/reaction-button";
 import { createSegmentedControlHandler } from "./handlers/segmented-control";
 import { createSelectBoxGroupHandler, createSelectBoxHandler } from "./handlers/select-box";
 import { createSkeletonHandler } from "./handlers/skeleton";
 import { createSnackbarHandler } from "./handlers/snackbar";
 import { createSwitchHandler } from "./handlers/switch";
+import { createTabsHandler } from "@/codegen/targets/react/component/handlers/tabs";
 import { createTextFieldHandler } from "./handlers/text-field";
 import { createToggleButtonHandler } from "./handlers/toggle-button";
+import { createListItemHandler } from "@/codegen/targets/react/component/handlers/list-item";
 
 export type { ComponentHandlerDeps };
 export type UnboundComponentHandler<T extends NormalizedInstanceNode["componentProperties"]> = (
@@ -59,6 +64,7 @@ export const unboundSeedComponentHandlers: Array<UnboundComponentHandler<any>> =
   createBottomSheetHandler,
   createCalloutHandler,
   createCheckboxHandler,
+  createCheckmarkHandler,
   createChipHandler,
   createContextualFloatingButtonHandler,
   createDividerHandler,
@@ -66,12 +72,15 @@ export const unboundSeedComponentHandlers: Array<UnboundComponentHandler<any>> =
   createFloatingActionButtonHandler,
   createHelpBubbleHandler,
   createIdentityPlaceholderHandler,
-  createPageBannerHandler,
+  createListItemHandler,
   createMannerTempBadgeHandler,
   createMannerTempHandler,
   createMenuSheetHandler,
   createMultilineTextFieldHandler,
+  createPageBannerHandler,
   createProgressCircleHandler,
+  createRadioGroupItemHandler,
+  createRadioMarkHandler,
   createReactionButtonHandler,
   createSegmentedControlHandler,
   createSelectBoxGroupHandler,
@@ -79,6 +88,7 @@ export const unboundSeedComponentHandlers: Array<UnboundComponentHandler<any>> =
   createSkeletonHandler,
   createSnackbarHandler,
   createSwitchHandler,
+  createTabsHandler,
   createTextFieldHandler,
   createToggleButtonHandler,
 ];