+++ packages/figma/src/entities/data/__generated__/component-sets/checkmark.mjs
@@ -2,18 +2,25 @@ export const metadata = {
   "name": "🟢 Checkmark",
   "key": "fae60fb392f55bde60de1dbccb2f453320068805",
   "componentPropertyDefinitions": {
-    "Shape": {
-      "type": "VARIANT",
-      "variantOptions": [
-        "Square",
-        "Ghost"
-      ]
-    },
     "Size": {
       "type": "VARIANT",
       "variantOptions": [
         "Medium",
         "Large"
+      ]
+    },
+    "Tone": {
+      "type": "VARIANT",
+      "variantOptions": [
+        "🚫[Deprecated]Brand",
+        "Neutral"
+      ]
+    },
+    "Shape": {
+      "type": "VARIANT",
+      "variantOptions": [
+        "Square",
+        "Ghost"
       ]
     },
     "Selected": {