+++ packages/figma/src/entities/data/__generated__/component-sets/divider.mjs
@@ -8,8 +8,8 @@ export const metadata = {
     "Tone": {
       "type": "VARIANT",
       "variantOptions": [
-        "Neutral",
-        "Neutral Muted"
+        "Neutral Muted",
+        "Neutral Subtle"
       ]
     },
     "Orientation": {