+++ packages/qvism-preset/src/vars/component/radio.mjs
@@ -14,14 +14,14 @@ export const vars = {
       }
     }
   },
-  "weightDefault": {
+  "weightRegular": {
     "enabled": {
       "label": {
         "fontWeight": "var(--seed-font-weight-regular)"
       }
     }
   },
-  "weightStronger": {
+  "weightBold": {
     "enabled": {
       "label": {
         "fontWeight": "var(--seed-font-weight-bold)"