+++ tools/figma-v3-migration/src/main/data/__generated__/v3-component-sets/list-header.d.ts
@@ -1,0 +1,27 @@
+export declare const metadata: {
+  "name": "🟢 List Header",
+  "key": "609f93ed0608ef0a6d9a351e47595ad631bae0fa",
+  "componentPropertyDefinitions": {
+    "Title#28588:0": {
+      "type": "TEXT",
+      "defaultValue": "리스트 그룹 제목"
+    },
+    "Suffix": {
+      "type": "VARIANT",
+      "defaultValue": "None",
+      "variantOptions": [
+        "None",
+        "Button",
+        "Custom"
+      ]
+    },
+    "Title Weight": {
+      "type": "VARIANT",
+      "defaultValue": "Medium",
+      "variantOptions": [
+        "Medium",
+        "Bold"
+      ]
+    }
+  }
+};