+++ tools/figma-v3-migration/src/main/data/__generated__/v3-component-sets/list-item-suffix-icon-button.mjs
@@ -1,0 +1,15 @@
+export const metadata = {
+  "name": "List Item/Suffix/Icon Button",
+  "key": "5636566f6de6f58200dce388f7b1ac9f517b30e1",
+  "componentPropertyDefinitions": {
+    "Number of Button": {
+      "type": "VARIANT",
+      "defaultValue": "1",
+      "variantOptions": [
+        "1",
+        "2",
+        "3"
+      ]
+    }
+  }
+};