+++ tools/figma-v3-migration/src/main/data/__generated__/v3-component-sets/resizable-child.d.ts
@@ -1,0 +1,19 @@
+export declare const metadata: {
+  "name": "🟢 ResizableChild",
+  "key": "df43b92004c850e7c6d2869e7a4ba1ef9a2d7db6",
+  "componentPropertyDefinitions": {
+    "Size": {
+      "type": "VARIANT",
+      "defaultValue": "12",
+      "variantOptions": [
+        "12",
+        "14",
+        "16",
+        "20",
+        "24",
+        "28",
+        "32"
+      ]
+    }
+  }
+};